+++ bluemart/MainViews/FavoritesPage.xaml.cs
@@ -85,7 +85,7 @@ namespace bluemart.MainViews
 			foreach (string productID in mFavoritesModel.GetProductIDs()) {
 				string ImagePath = ProductModel.mRootFolderPath + "/" + ParseConstants.IMAGE_FOLDER_NAME + "/" + ProductModel.mProductImageNameDictionary[productID] + ".jpg";
 				string ProductName = ProductModel.mProductNameDictionary [productID];
-				double price = ProductModel.mProductPriceDictionary [productID];
+				decimal price = ProductModel.mProductPriceDictionary [productID];
 				string quantity = ProductModel.mProductQuantityDictionary [productID];
 				mProductList.Add (new Product (productID,ProductName, ImagePath, price, quantity)); 
 			}
