+++ bluemart/Models/Remote/ProductModel.cs
@@ -18,7 +18,7 @@ namespace bluemart.Models.Remote
 		public static Dictionary<string,string> mProductNameDictionary;
 		public static Dictionary<string,string> mProductImageIDDictionary;
 		public static Dictionary<string,string> mProductImageNameDictionary;
-		public static Dictionary<string,double> mProductPriceDictionary;
+		public static Dictionary<string,decimal> mProductPriceDictionary;
 		public static Dictionary<string,string> mProductQuantityDictionary;
 		public static Dictionary<string,HashSet<string>> mProductCategoryIDDictionary;
 		private static FavoritesClass mFavoritesModel = new FavoritesClass();
@@ -40,7 +40,7 @@ namespace bluemart.Models.Remote
 			mProductNameDictionary = new Dictionary<string, string> ();
 			mProductImageIDDictionary = new Dictionary<string, string> ();
 			mProductImageNameDictionary = new Dictionary<string, string> ();
-			mProductPriceDictionary = new Dictionary<string, double> ();
+			mProductPriceDictionary = new Dictionary<string, decimal> ();
 			mProductQuantityDictionary = new Dictionary<string, string> ();
 			mProductCategoryIDDictionary = new Dictionary<string,HashSet<string>> ();
 		}
@@ -73,7 +73,7 @@ namespace bluemart.Models.Remote
 				tempProduct.ImageName = productObject.Get<string> (ParseConstants.PRODUCT_ATTRIBUTE_IMAGENAME);
 				tempProduct.ImageID = productObject.Get<string> (ParseConstants.PRODUCT_ATTRIBUTE_IMAGEID);
 				tempProduct.CategoryId = productObject.Get<string> (ParseConstants.PRODUCT_ATTRIBUTE_CATEGORYID);
-				tempProduct.Price = productObject.Get<double> (ParseConstants.PRODUCT_ATTRIBUTE_PRICE);
+				tempProduct.Price = productObject.Get<decimal> (ParseConstants.PRODUCT_ATTRIBUTE_PRICE);
 				tempProduct.Quantity = productObject.Get<string> (ParseConstants.PRODUCT_ATTRIBUTE_QUANTITY);
 
 				tempList.Add (tempProduct);
@@ -88,6 +88,7 @@ namespace bluemart.Models.Remote
 				DateTime? localUpdate = mUserClass.GetProductsUpdatedDateFromUser ();
 				var query = ParseObject.GetQuery (ParseConstants.PRODUCTS_CLASS_NAME).OrderByDescending (ParseConstants.UPDATEDATE_NAME).Limit (1);
 				var parseObject = query.FirstAsync ().Result;
+				
 				DateTime? remoteUpdate = parseObject.UpdatedAt;
 				//update category class
 				if (remoteUpdate > localUpdate) {
